updating run to try pulling bundle multiple times before failing. returning status from runner

+++ scripts/icetray/step1/run_step1.py
@@ -6,6 +6,7 @@ import os
 import zipfile
 import shutil
 import tempfile
+import time
 
 from pathlib import Path, PosixPath
 import concurrent.futures
@@ -44,16 +45,19 @@ def get_logfilenames(infile: Path, outdir: Path) -> tuple[Path, Path]:
     infilename = str(remove_extension(infile))
     infilenwords = infilename.split('_')
     stdoutfilename = "_".join(
-        ["LOG", "OUT", "Pass3", "Step1"] + infilenwords[1:]) + ".out"
+        ["LOG", "Pass3", "Step1"] + infilenwords[1:]) + ".out"
     stderrfilename = "_".join(
-        ["LOG", "ERR", "Pass3", "Step1"] + infilenwords[1:]) + ".err"
+        ["LOG", "Pass3", "Step1"] + infilenwords[1:]) + ".err"
     return outdir / stdoutfilename, outdir / stderrfilename
 
 def generate_command(scriptloc: Path,
                      infile: Path,
                      gcd: Path,
-                     outfile: Path) -> str:
-    command = "python3 " + scriptloc + f" -i {infile} -g {gcd} -o {outfile} --qify"
+                     outfile: Path,
+                     qify: bool = False) -> str:
+    command = f"python3 {scriptloc} -i {infile} -g {gcd} -o {outfile}"
+    if qify:
+        command += " --qify"
     return command
 
 # Taken from LTA
@@ -69,17 +73,45 @@ def get_sha512sum(filename: Union[str, Path]) -> str:
             h.update(mv[:n])
     return h.hexdigest()
 
-def prepare_inputs(outdir: Path, bundle: Path, checksum: str, gcddir: Path):
+def get_bundle(bundle: Path, outdir: Path, retry_attempts: int = 5):
+    for a in range(retry_attempts):
+        try:
+            subprocess.run(f"scp ranch.tacc.utexas.edu:{bundle} {str(outdir) + '/'}", shell=True)
+            print(f"Successfully retrieved bundle: {bundle}")
+            break
+        except subprocess.CalledProcessError as e:
+            print(f"Retrieving bundled {bundle}  failed (attempt {a + 1}/{retry_attempts})")
+            print(f"Error output: {e.stderr.strip()}")
+            if a < (retry_attempts - 1):
+                print(f"Waiting 10 seconds")
+                time.sleep(10)
+            else:
+                raise
+
+def prepare_inputs(outdir: Path,
+                   scratchdir: Path,
+                   bundle: Path,
+                   checksum: str,
+                   gcddir: Path) -> list:
     if not outdir.exists():
         outdir.mkdir(parents=True, exist_ok=True)
 
-    if not (outdir / bundle.name).exists():
-        subprocess.run(f"scp {os.environ['ARCHIVER']}:{bundle} {str(outdir) + '/'}", shell=True)
-        bundle_sha512sum = get_sha512sum(outdir / bundle.name)
+    if (scratchdir / bundle.name).exists():
+        # Checking if available bundle is good
+        bundle_sha512sum = get_sha512sum(scratchdir / bundle.name)
         if bundle_sha512sum != checksum:
-            raise Exception(f"Bundle {bundle} checksum is not the same")
+            shutil.rmtree(scratchdir/ bundle.name)
+            get_bundle(bundle, scratchdir)
+            bundle_sha512sum = get_sha512sum(scratchdir / bundle.name)
+            if bundle_sha512sum != checksum:
+                # Bundle from tape is borked...
+                raise Exception(f"Bundle {bundle} checksum is not what we expect. Something wrong with tape bundle???")
+    else:
+        # Getting bundle if it doesn't exist
+        print(f"Getting bundle {bundle.name}")
+        get_bundle(bundle, scratchdir)
 
-    scratch_bundle_loc = outdir / bundle.name
+    scratch_bundle_loc = scratchdir / bundle.name
 
     infiles = [f for f in zipfile.ZipFile(scratch_bundle_loc).namelist()
                if ".tar.gz" in f]
@@ -96,10 +128,9 @@ def prepare_inputs(outdir: Path, bundle: Path, checksum: str, gcddir: Path):
          for f in infiles]
 
     return inputs
+def runner(infiles: tuple[Path, Path, Path, Path]) -> str:
 
-def runner(infiles: tuple[Path, Path, Path, Path]):
-    # Creating a dir on the local disk
-    tmpdir = Path(tempfile.mkdtemp(dir="/tmp"))
+    print(shutil.disk_usage("/tmp"))
 
     # Getting file and file paths
     gcddir = infiles[0]
@@ -107,15 +138,20 @@ def runner(infiles: tuple[Path, Path, Path, Path]):
     infile = infiles[2]
     outdir = infiles[3]
 
+    # Creating a temporary working dir for this instance
+    tmpdir = Path(tempfile.mkdtemp(dir=str(bundle.parent)))
+
     gcd = get_gcd(infile, gcddir)
     if not gcd.exists():
         raise FileNotFoundError("No GCD")
 
+    print(f"Copying GCD file: {gcd}")
     shutil.copyfile(gcd, tmpdir / gcd.name)
     local_gcd = tmpdir / gcd.name
 
     # Prepping files and file paths
     ## Extracting in file from bundle
+    print(f"Extracting {infile}")
     zipfile.ZipFile(bundle).extract(str(infile), path=tmpdir)
     local_infile = tmpdir / infile
     if not local_infile.exists():
@@ -127,7 +163,7 @@ def runner(infiles: tuple[Path, Path, Path, Path]):
     outfile = outdir / outfilename
 
     if outfile.exists():
-        raise Exception("Output file already exists")
+        return f"Output file {outfile} already exists"
 
     local_stdout_file, local_stderr_file = get_logfilenames(infile,
                                                             tmpdir)
@@ -138,62 +174,73 @@ def runner(infiles: tuple[Path, Path, Path, Path]):
         Path("/opt/pass3/scripts/icetray/step1/pass3_reprocess_PFRaw.py"),
         local_infile,
         local_gcd,
-        local_temp_outfile)
+        local_outfile,
+        qify = True)
     moni_command = generate_command(
-        Path("/opt/pass3/scripts/icetray/step1/pass3_check_charge_filter.py"),
-        local_temp_outfile,
+        Path("/opt/pass3/scripts/icetray/step1/pass3_step1_unpackdst.py"),# pass3_check_charge_filter.py"),
+        local_outfile,
         local_gcd,
         local_outfile)
 
     # run step 1
-    # We are first running the online processing that is the same as done 
-    # at the south pole. we then read the file back in, rehydrate it, and 
+    # We are first running the online processing that is the same as done
+    # at the south pole. we then read the file back in, rehydrate it, and
     # run some moni code on it to make sure we are doing the right thing.
     try:
         with open(local_stdout_file, "w") as stdout, open(local_stderr_file, "w") as stderr:
             # TODO: Do we need to check the GCD file?
             stdout.write(
                     f"Start Time: {datetime.datetime.now(datetime.timezone.utc)}\n")
-            stdout.write(f"Hostname: {os.environ['HOSTNAME']}")
+            stdout.write(f"Hostname: {os.environ['HOSTNAME']}\n")
             try:
                 subprocess.run(command, shell=True, stdout=stdout, stderr=stderr)
             except:
-                raise Exception(
-                    f"ALERT: {infile} in {bundle} has failed to process")
+                return f"ALERT: {infile} in {bundle} has failed to process\n"
             stdout.write(
                     f"End Time PFRAW: {datetime.datetime.now(datetime.timezone.utc)}\n")
             try:
                 subprocess.run(
                     moni_command, shell=True, stdout=stdout, stderr=stderr)
             except:
-                raise Exception(
-                    f"ALERT: {infile} in {bundle} has failed during moni")      
+                return f"ALERT: {infile} in {bundle} has failed during moni\n"
             stdout.write(
                     f"End Time: {datetime.datetime.now(datetime.timezone.utc)}\n")
     finally:
+        print("Copying logs")
         shutil.copyfile(local_stdout_file, stdout_file)
         shutil.copyfile(local_stderr_file, stderr_file)
 
     # create checksum of output file
+    print(f"Getting sha512sum for {local_outfile}")
     sha512sum = get_sha512sum(local_outfile)
     print(f"file: {local_outfile} sha512sum: {sha512sum}")
-    with Path.open(outfile + ".sha512sum", "w"):
-        file.write(f"{sha512sum}")
+    with Path.open(str(outfile) + ".sha512sum", "w") as file:
+        file.write(f"{outfile} {sha512sum}")
 
     # Copying from local dir to absolute dir
+    print("Copying output file")
     shutil.copyfile(local_outfile, outfile)
+
+    sha512sum_final = get_sha512sum(outfile)
+    if sha512sum_final != sha512sum:
+        raise Exception(f"Copying file {local_outfile} from tmp storage to final storage failed.")
+
     # TODO: move charge and filter rate file
-    shutil.copyfile(local_outfile + ".npz", outfile + ".npz",)
-    shutil.copyfile(local_outfile + ".txt", outfile + ".txt")
+    print("Copying moni files")
+    shutil.copyfile(str(local_outfile) + ".npz", str(outfile) + ".npz",)
+    shutil.copyfile(str(local_outfile) + ".txt", str(outfile) + ".txt")
 
     shutil.rmtree(tmpdir)
+    return f"Processing file {infile} from bundle {bundle} was SUCCESSFUL. Output file {outfile} with checksum {sha512sum}."
 
 def run_parallel(infiles, max_num=1):
+    print(f"max workers: {max_num}")
+    print(f"length infiles: {len(infiles)}")
     with concurrent.futures.ProcessPoolExecutor(
         max_workers = max_num) as executor:
         futures = executor.map(runner, infiles)
         for f in futures:
-            print(f.result())
+            print(f)
     return futures
 
 if __name__ == "__main__":
@@ -203,13 +250,17 @@ if __name__ == "__main__":
                         type=Path,
                         required=True)
     parser.add_argument("--bundle",
-                        help="path to bundle on ranch", 
+                        help="path to bundle on ranch",
                         type=Path,
                         required=True)
     parser.add_argument("--outdir",
                         help="",
                         type=Path,
                         required=True)
+    parser.add_argument("--scratchdir",
+                        help="Path where work should be done",
+                        type=Path,
+                        default="/tmp")
     parser.add_argument("--checksum",
                         help="bundle sha512sum",
                         type=str,
@@ -221,11 +272,18 @@ if __name__ == "__main__":
     args=parser.parse_args()
 
     if args.maxnumcpus == 0:
-        numcpus =  None
+        numcpus = os.cpu_count()
     else:
         numcpus = args.maxnumcpus
 
+    # print(os.environ)
+    # os.environ['OPENBLAS_MAIN_FREE'] = str(1)
+    # os.system(f'taskset -cp 0-{numcpus} {os.getpid()}')
+
+    print(f"CPU count: {numcpus}")
+
     inputs = prepare_inputs(args.outdir,
+                            args.scratchdir,
                             args.bundle,
                             args.checksum,
                             args.gcddir)
